+++ lint.sh
@@ -1,5 +1,7 @@
+#!/bin/bash
+
 cd node/
 [ -d node_modules ] && rm -rf node_modules
 yarn cache clean
-yarn --frozen-lock
+yarn --frozen-lockfile
 yarn lint