+++ node/resolvers/document/index.ts
@@ -1,4 +1,5 @@
-import { union } from 'ramda'
+import { prop, union } from 'ramda'
+import {parseFieldsToJson }  from '../../utils'
 import { mapKeyValues } from '../../utils/object'
 
 export const queries = {
@@ -29,7 +30,8 @@ export const mutations = {
 
   updateDocument: async (_, args, { dataSources: { document } }) => {
     const { acronym, document: { fields } } = args
-    const { Id, Href, DocumentId } = await document.updateDocument(acronym, fields)
+    const id = prop('id', parseFieldsToJson(fields))
+    const { Id, Href, DocumentId } = await document.updateDocument(acronym, id, fields)
     return { cacheId: DocumentId, id: Id, href: Href, documentId: DocumentId }
   },